fix: added locking to cart mutating routes to avoid race conditions (#11229)

+++ src/Core/Checkout/Cart/SalesChannel/CartOrderRoute.php
@@ -7,6 +7,7 @@ use Shopware\Core\Checkout\Cart\Cart;
 use Shopware\Core\Checkout\Cart\CartCalculator;
 use Shopware\Core\Checkout\Cart\CartContextHasher;
 use Shopware\Core\Checkout\Cart\CartException;
+use Shopware\Core\Checkout\Cart\CartLocker;
 use Shopware\Core\Checkout\Cart\Event\CheckoutOrderPlacedCriteriaEvent;
 use Shopware\Core\Checkout\Cart\Event\CheckoutOrderPlacedEvent;
 use Shopware\Core\Checkout\Cart\Order\OrderPersisterInterface;
@@ -28,7 +29,6 @@ use Shopware\Core\Framework\Validation\DataBag\RequestDataBag;
 use Shopware\Core\Profiling\Profiler;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\Lock\LockFactory;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
@@ -36,8 +36,6 @@ use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 #[Package('checkout')]
 class CartOrderRoute extends AbstractCartOrderRoute
 {
-    private const LOCK_TTL = 30;
-
     /**
      * @internal
      */
@@ -52,7 +50,8 @@ class CartOrderRoute extends AbstractCartOrderRoute
         private readonly TaxProviderProcessor $taxProviderProcessor,
         private readonly AbstractCheckoutGatewayRoute $checkoutGatewayRoute,
         private readonly CartContextHasher $cartContextHasher,
-        private readonly LockFactory $lockFactory,
+        private readonly ExtensionDispatcher $extensions,
+        private readonly CartLocker $cartLocker
     ) {
     }
 
@@ -70,28 +69,19 @@ class CartOrderRoute extends AbstractCartOrderRoute
             throw CartException::hashMismatch($cart->getToken());
         }
 
-        // we use this state in stock updater class, to prevent duplicate available stock updates
-        $context->addState('checkout-order-route');
-
-        $lock = $this->lockFactory->createLock('cart-order-route-' . $cart->getToken(), self::LOCK_TTL);
-        if (!$lock->acquire()) {
-            throw CartException::cartLocked($cart->getToken());
-        }
-
-        try {
-            $calculatedCart = $this->cartCalculator->calculate($cart, $context);
+        return $this->cartLocker->locked($cart->getToken(), function () use ($cart, $context, $data) {
+            // we use this state in stock updater class, to prevent duplicate available stock updates
+            $context->addState('checkout-order-route');
 
             $response = $this->checkoutGatewayRoute->load(new Request($data->all(), $data->all()), $cart, $context);
             $calculatedCart->addErrors(...$response->getErrors());
 
             $this->taxProviderProcessor->process($calculatedCart, $context);
 
-            $this->addCustomerComment($calculatedCart, $data);
-            $this->addAffiliateTracking($calculatedCart, $data);
-
-            $preOrderPayment = Profiler::trace('checkout-order::pre-payment', fn () => $this->paymentProcessor->validate($calculatedCart, $data, $context));
-
-            $orderId = Profiler::trace('checkout-order::order-persist', fn () => $this->orderPersister->persist($calculatedCart, $context));
+            if (Feature::isActive('v6.8.0.0')) {
+                // @deprecated tag:v6.8.0 - After the cart is deleted, the lock is no longer needed. The following operations should be moved outside the locked closure.
+                $this->cartPersister->delete($context->getToken(), $context);
+            }
 
             $criteria = new Criteria([$orderId]);
             $criteria
@@ -127,23 +117,13 @@ class CartOrderRoute extends AbstractCartOrderRoute
                 $this->eventDispatcher->dispatch($event);
             });
 
-            $this->cartPersister->delete($context->getToken(), $context);
+            if (!Feature::isActive('v6.8.0.0')) {
+                // cart will delete immediately after order is created to avoid inconsistencies.
+                $this->cartPersister->delete($context->getToken(), $context);
+            }
 
-            // @deprecated tag:v6.7.0 - remove post payment completely
-            Feature::callSilentIfInactive('v6.7.0.0', function () use ($orderEntity, $data, $context, $orderId, $preOrderPayment): void {
-                try {
-                    Profiler::trace('checkout-order::post-payment', function () use ($orderEntity, $data, $context, $preOrderPayment): void {
-                        $this->preparedPaymentService->handlePostOrderPayment($orderEntity, $data, $context, $preOrderPayment);
-                    });
-                } catch (PaymentException) {
-                    throw CartException::invalidPaymentButOrderStored($orderId);
-                }
-            });
-        } finally {
-            $lock->release();
-        }
-
-        return new CartOrderRouteResponse($orderEntity);
+            return new CartOrderRouteResponse($orderEntity);
+        });
     }
 
     private function addCustomerComment(Cart $cart, DataBag $data): void
